MDEV-22430 Assertion ha_info->ht() != binlog_hton' failed in MYSQL_BIN_LOG::unlog_xa_prepare MDEV-22757 Assertion !binlog || exist_hton_without_prepare' failed in MYSQL_BIN_LOG::unlog_xa_prepare
Asserts (actually the same one just in different version of sources) fired
having not capture two more not apparent possiblities. They are
masked out hton error out of REPLACE execution (so at later xa-prepare
that engine is still present as read-write) and a prepare-capable engine
which also may not be an actual participant in the xa transation. That
engine, such as SEQUENCE, though does create its own event block.

It's rather hard to embed those