MDEV-17066: Bytes lost or Assertion `status_var.local_memory_used == 0 after DELETE with subquery with ROLLUP
The issue here is when records are read from the temp table via a
cache(rr_from_cache), the cache is allocated for each execution of
the correlated subquery. But the deallocation of the cache only
happened after the entire execution of the parent select was done.
The fix here would be free the cache on each execution of the
subquery.